+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - save and restore grouped sessions (used with multi-monitor workflow)
 - save and restore active and alternate windows in grouped sessions
 - if there are no grouped sessions, do not output empty line to "last" file
+- restore active and alternate windows only if they are present in the "last" file
 
 ### v1.5.0, 2014-11-09
 - add support for restoring neovim sessions